\name{permimp-package}
\alias{permimp-package}
\docType{package}
\title{
\packageTitle{permimp}
}
\description{
\packageDescription{permimp}
}
\details{

\packageIndices{permimp}
}
\author{
Maintainer: Dries Debeer \email{dries.debeer@uzh.ch}

Authors:
\itemize{
  \item Carolin Strobl
  \item Torsten Hothorn
}
}
\keyword{ package }

